Renamed AppLayout to VirtoAppLayout
